refactor(graph): Add OpType and Attributes to Node interface
This commit extends the core graph.Node interface to include OpType() and Attributes() methods. This is a necessary refactoring to allow the model exporter to query nodes for their type and parameters, enabling serialization to formats like ZMF.

Updates the layers in the activations package to implement the new interface methods.